# PoC: assume code ends with STOP