--- a/lib/evmone/baseline.cpp
+++ b/lib/evmone/baseline.cpp
@@ -112,7 +112,7 @@ evmc_result baseline_execute(ExecutionState& state) noexcept
 
     const auto code_end = code + code_size;
     auto* pc = code;
-    while (pc != code_end)
+    while (true)  // Guaranteed to terminate because code must end with STOP.
     {
         const auto op = *pc;