@chfast 2021-06-01
This optimization is enabled in #315 by copying and padding the original code.